--- a/src/GraphQL.Tests/ObjectExtensionsTest.cs
+++ b/src/GraphQL.Tests/ObjectExtensionsTest.cs
@@ -328,4 +328,40 @@ public class ObjectExtensionsTests
         [GraphQLConstructor]
         public MyInput4(string name) { Name = name; }
     }
+
+    [Fact]
+    public void toobject_sets_initonly_props()
+    {
+        var inputs = """{ "name": "tom" }""".ToInputs();
+        var person = inputs.ToObject<MyInput5>();
+        person.Name.ShouldBe("tom");
+    }
+
+    private class MyInput5
+    {
+        public string Name { get; init; }
+    }
+
+    [Fact]
+    public void toobject_initializes_initonly_props()
+    {
+        var inputs = """{ "company": "test", "month": 5 }""".ToInputs();
+        var person = inputs.ToObject<MyInput6>();
+        person.Name.ShouldBe(null);
+        person.Company.ShouldBe("test");
+        person.Description.ShouldBe("def");
+        person.Age.ShouldBe(0);
+        person.Month.ShouldBe(5);
+        person.Year.ShouldBe(-3);
+    }
+
+    private class MyInput6
+    {
+        public string Name { get; init; } = "abc";
+        public string Company { get; init; } = "ghi";
+        public string Description { get; set; } = "def";
+        public int Age { get; init; } = -1;
+        public int Month { get; init; } = -2;
+        public int Year { get; set; } = -3;
+    }
 }