… docs2/site/docs/migrations/migration8.md
… src/GraphQL.ApiTests/net50/GraphQL.approved.txt
… src/GraphQL.ApiTests/net60/GraphQL.approved.txt
… src/GraphQL.Benchmarks/Benchmarks/ExecutionBenchmark.cs
… src/GraphQL.Benchmarks/Benchmarks/ToObjectBenchmark.cs
… src/GraphQL.Benchmarks/Properties/launchSettings.json
… src/GraphQL/Types/Composite/InputObjectGraphType.cs