This PR makes `ToObject` use a separate function to pull reflection information to determine 'how' to construct the object.  This makes `ToObject` itself much simpler.

There are a couple implementation changes, but otherwise is functionally identical:
- Private setters for public properties are not supported
- Writing to read-only fields is not supported
- Init-only setters are always initialized to their default value if not specified

This allows for:
- Caching of reflection information independent of `ToObject` operation (separate PR)
- Re-use of reflection information for dynamically-compiled `ToObject` implementations (separate PR)
- 100% compatible with a source-generated `ToObject` implementation

Also todo in separate PR:
- Clean up logic in `ObjectExtensions.GetPropertyValue`

Note:
- Does not set init-only properties first, as would occur with a source-generated implementation
- We could set ALL undefined properties to their default values, not just init-only ones, making reflection, source-generated and dynamically-compiled versions simpler.