@Shane32 2023-12-03
Test scenario:

```cs
    private IGraphType _humanInputType;
    private Dictionary<string, object> _dic;

    [GlobalSetup]
    public void GlobalSetup()
    {
        var services = new ServiceCollection();
        services.AddGraphQL(b => b
            .AddSchema<StarWarsSchema>()
            .AddGraphTypes(typeof(StarWarsQuery).Assembly));
        services.AddSingleton<StarWarsData>();

        _provider = services.BuildServiceProvider();
        _schema = _provider.GetRequiredService<ISchema>();
        _schema.Initialize();
        _executer = new DocumentExecuter();
        _cachedExecuter = new DocumentExecuter(new GraphQLDocumentBuilder(), new DocumentValidator(), new DefaultExecutionStrategySelector(), new[] { new MemoryDocumentCache() });
        _humanInputType = _schema.AllTypes["HumanInput"];
        _dic = new Dictionary<string, object>()
        {
            { "name", "Test" },
            { "homePlanet", "Abc" }
        };
    }

    [Benchmark]
    public void ToObject()
    {
        _ = (Human)ObjectExtensions.ToObject(_dic, typeof(Human), _humanInputType);
    }
```

BEFORE

| Method   | Mean     | Error   | StdDev   | Gen0   | Allocated |
|--------- |---------:|--------:|---------:|-------:|----------:|
| ToObject | 409.2 ns | 8.18 ns | 11.73 ns | 0.0482 |     304 B |

AFTER

| Method   | Mean     | Error    | StdDev   | Gen0   | Allocated |
|--------- |---------:|---------:|---------:|-------:|----------:|
| ToObject | 356.7 ns |  4.48 ns |  4.19 ns | 0.0353 |     224 B |