@gao-artur 2023-12-13
> In the sample you provided, it would be more likely that the field would be required and have a default value of an empty list.

Then, to avoid confusion, we must have an analyzer to ensure "legal" defaults are applied on graph and source types. Please correct me if this table is wrong:
|Graph type field|Source type member default|
|-|-|
| Nullable without default | Nothing or null |
| Nullable with default | Nothing or same as on graph |
| Not nullable without default | Nothing |
| Not nullable with default | Nothing or same as on graph |

If we create such an analyzer, then we can implement this note:

> We could set ALL undefined properties to their default values, not just init-only ones, making reflection, source-generated and dynamically-compiled versions simpler.

Because there won't be runtime surprises for the developer.